pyproject.toml
[project]
name = "shellclaw"
version = "0.1.0"
description = "Hardware-agnostic OpenClaw agent orchestration CLI"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "typer[all]>=0.15.0",
    "pyyaml>=6.0",
]

[project.scripts]
shellclaw = "shellclaw.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/shellclaw"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "ruff>=0.9.0",
    "mypy>=1.14",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests (require Docker + OpenShell)",
]

[tool.coverage.run]
source = ["shellclaw"]

[tool.coverage.report]
fail_under = 80

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
